refactor: Pull out lsp related code into library file
